+++ NEWS.md
@@ -1,5 +1,9 @@
 # sooty dev
 
+* dataset() is now deprecated in favour of datasource()
+ 
+* Created standalone function available_datasets(), and removed `@curated`, `@available_datasets` from the datasource object. 
+
 * Now using more robust model of GDAL protocol and absolute URLs. 
 
 * Added an S7 object to set dataset id and retrieve files.